+++ integrations/slack/src/actions/index.ts
@@ -1,1 +1,4 @@
 export * from './queryAskAI';
+export * from './ingestConversation';
+export * from './inferUserIntent';
+export * from './types';