+++ integrations/github/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @gitbook/integration-github
 
+## 0.6.6
+
+### Patch Changes
+
+- c770fee: Fix proxy for the webhook endpoint
+
+## 0.6.5
+
+### Patch Changes
+
+- 247db36: Improve support for proxy
+- Updated dependencies [7cb9bbc]
+    - @gitbook/api@0.152.1
+
+## 0.6.4
+
+### Patch Changes
+
+- d7fa48f: Replace custom code for tasks to leverage the new runtime API.
+- Updated dependencies [d7fa48f]
+    - @gitbook/runtime@0.23.0
+
 ## 0.6.3
 
 ### Patch Changes